finished climber
use same motion magic logic as elevator, but gravity type does not exist. can use arm-cosine theoretically but since it is being used to climb and the robot is the thing moving instead of the arm, the gravity profile is different so i dont know how accurate it would be. so potential testing